+++ projects/app-lob/src/app/grid-finjs-dock-manager/grid-finjs-dock-manager.component.scss
