+++ projects/app-lob/src/app/index/index.component.ts
@@ -3,7 +3,7 @@ import { NavigationStart, Route, Router } from "@angular/router";
 import { IgxNavigationDrawerComponent } from "igniteui-angular";
 import { filter } from "rxjs/operators";
 import { gridDynamicChartRoutesData } from "../grid-dynamic-chart-data/grid-dynamic-chart-data-routes-data";
-import { finJsDockManagerGridRoutes } from "../grid-finjs-dockManager/grid-finjs-dock-manager-routes-data";
+import { finJsDockManagerGridRoutes } from "../grid-finjs-dock-manager/grid-finjs-dock-manager-routes-data";
 import { gridfinjsRoutesData} from "../grid-finjs/grid-finjs-routes-data";
 import { gridsRoutesData } from "../grid/grid-routes-data";
 import { hierarchicalGridRoutesData } from "../hierarchical-grid/hierarchical-grid-routes-data";
